requirements.txt
feedparser
requests
google-generativeai
python-dateutil